# public/swagger.yml
...
swagger: '2.0'
...
info:
  version: "1"
  title: Clowder API
  description: |
    Documentation for the [Clowder](http://clowder.ncsa.illinois.edu/) API

################################################################################
#                  Host, Base Path, Schemes and Content Types                  #
################################################################################
host: clowder.ncsa.illinois.edu
basePath: /clowder/api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json

################################################################################
#                                   Tags                                       #
################################################################################
...
securityDefinitions:
  login:
    type: basic
  key:
    type: apiKey
    name: key
    in: query
security:
  - key: []
  - login: []

################################################################################
# Parameters                                                                   #
################################################################################

################################################################################
# Paths                                                                        #
################################################################################
paths:
...
  /files/{id}/previews/{p_id}:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: id
          in: path
        - required: true
          type: string
          name: p_id
          in: path
      tags:
        - files
      summary: Attach existing preview to file
      description: A file is the raw bytes plus metadata.
...
  /collections/{c_id}/previews/{p_id}:
    post:
      responses:
        200:
          description: OK
      parameters:
        - required: true
          type: string
          name: c_id
          in: path
        - required: true
          type: string
          name: p_id
          in: path
      tags:
        - collections
      summary: Attach existing preview to collection
      description: Collections are groupings of datasets
...